[Storage Service] Split up logic into different files/modules.
