+++ state-sync/aptos-data-client/src/aptosnet/logging.rs
@@ -5,7 +5,7 @@ use crate::Error;
 use aptos_config::network_id::PeerNetworkId;
 use aptos_logger::Schema;
 use serde::Serialize;
-use storage_service_types::StorageServiceRequest;
+use storage_service_types::requests::StorageServiceRequest;
 
 #[derive(Schema)]
 pub struct LogSchema<'a> {
