+++ state-sync/aptos-data-client/src/aptosnet/mod.rs
@@ -36,12 +36,13 @@ use network::{
 use rand::seq::SliceRandom;
 use std::{convert::TryFrom, fmt, sync::Arc, time::Duration};
 use storage_service_client::StorageServiceClient;
-use storage_service_types::{
-    Epoch, EpochEndingLedgerInfoRequest, NewTransactionOutputsWithProofRequest,
-    NewTransactionsWithProofRequest, StateValuesWithProofRequest, StorageServerSummary,
-    StorageServiceRequest, StorageServiceResponse, TransactionOutputsWithProofRequest,
-    TransactionsWithProofRequest,
+use storage_service_types::requests::{
+    EpochEndingLedgerInfoRequest, NewTransactionOutputsWithProofRequest,
+    NewTransactionsWithProofRequest, StateValuesWithProofRequest, StorageServiceRequest,
+    TransactionOutputsWithProofRequest, TransactionsWithProofRequest,
 };
+use storage_service_types::responses::{StorageServerSummary, StorageServiceResponse};
+use storage_service_types::Epoch;
 use tokio::{runtime::Handle, task::JoinHandle};
 
 mod logging;
