+++ state-sync/aptos-data-client/src/aptosnet/state.rs
@@ -18,7 +18,8 @@ use std::{
     collections::{HashMap, HashSet},
     sync::Arc,
 };
-use storage_service_types::{StorageServerSummary, StorageServiceRequest};
+use storage_service_types::requests::StorageServiceRequest;
+use storage_service_types::responses::StorageServerSummary;
 
 /// Scores for peer rankings based on preferences and behavior.
 const MAX_SCORE: f64 = 100.0;
